moved Correlations plugin to Data
