--- src/plugins/correlations/opensearch_dashboards.json
@@ -1,8 +1,0 @@
-{
-  "id": "correlations",
-  "version": "opensearchDashboards",
-  "server": true,
-  "ui": true,
-  "requiredPlugins": ["savedObjects"],
-  "optionalPlugins": []
-}
