--- src/plugins/correlations/server/saved_objects/index.ts
@@ -1,6 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export { correlationsSavedObjectType } from './correlations';
