--- src/plugins/correlations/server/types.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface CorrelationsSetup {}
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface CorrelationsStart {}
