+++ src/plugins/data/server/index.ts
@@ -312,3 +312,5 @@ export const config: PluginConfigDescriptor<ConfigSchema> = {
 };
 
 export type { IndexPatternsService } from './index_patterns';
+
+export { correlationsSavedObjectType } from './saved_objects';
